Folder structure changes & renaming related to LnAddress & fee widgets
