+++ lib/user_app.dart
@@ -19,7 +19,7 @@ import 'package:c_breez/routes/home/home_page.dart';
 import 'package:c_breez/routes/initial_walkthrough/initial_walkthrough.dart';
 import 'package:c_breez/routes/initial_walkthrough/mnemonics/enter_mnemonics_page.dart';
 import 'package:c_breez/routes/initial_walkthrough/mnemonics/mnemonics_confirmation_page.dart';
-import 'package:c_breez/routes/lightning_address/lightning_address_page.dart';
+import 'package:c_breez/routes/ln_address/ln_address_page.dart';
 import 'package:c_breez/routes/lsp/select_lsp_page.dart';
 import 'package:c_breez/routes/network/network_page.dart';
 import 'package:c_breez/routes/payment_options/payment_options_page.dart';
@@ -155,7 +155,7 @@ class UserApp extends StatelessWidget {
                                   );
                                 case '/lnurl_pay':
                                   return FadeInRoute(
-                                    builder: (_) => const LightningAddressPage(),
+                                    builder: (_) => const LnAddressPage(),
                                     settings: settings,
                                   );
                                 case '/get_refund':
